backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
bitarray==3.8.0
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
cytoolz==1.1.0
dj-database-url==3.1.0
Django==5.2.10
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
frozenlist==1.8.0
hexbytes==1.3.1
idna==3.11
kombu==5.6.2
multidict==6.7.0
packaging==25.0
parsimonious==0.10.0
pillow==12.1.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyunormalize==17.0.0
redis==7.1.0
regex==2025.11.3
requests==2.32.5
rlp==4.1.0
six==1.17.0
sqlparse==0.5.5
toolz==1.1.0
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
web3==7.14.0
websockets==15.0.1
yarl==1.22.0
